API Separate CwpSearchPageController and remove Object reference in CwpSearchEngine
